requirements.txt
# ── Assignment PDF Permitted Libraries (submitted code may ONLY import these) ──
torch>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
wandb>=0.16.0
albumentations>=1.3.1

# ── PyTorch ecosystem ──────────────────────────────────────────────────────────
torchvision>=0.16.0         # ONLY for initial dataset download — never imported in model files

# ── Image I/O ──────────────────────────────────────────────────────────────────
pillow>=9.5.0
opencv-python-headless>=4.8.0

# ── Training utilities (only used in train.py / notebooks, NOT submitted models) ─
tqdm>=4.66.0
rich>=13.0.0
requests>=2.31.0
scipy>=1.11.0

# ── Jupyter / notebook support ────────────────────────────────────────────────
ipykernel>=6.25.0
ipywidgets>=8.1.0